[INLONG-3516][Manager] rename the manager-client-test to manager-client-example
