+++ inlong-manager/manager-client-example/pom.xml
@@ -25,8 +25,8 @@
         <artifactId>inlong-manager</artifactId>
         <version>1.1.0-incubating-SNAPSHOT</version>
     </parent>
-    <artifactId>manager-client-test</artifactId>
-    <name>Apache InLong - Manager Client Test</name>
+    <artifactId>manager-client-example</artifactId>
+    <name>Apache InLong - Manager Client Example</name>
 
     <dependencies>
         <dependency>
