+++ inlong-manager/manager-client-example/src/test/java/org/apache/inlong/manager/client/Binlog2KafkaExample.java
@@ -18,7 +18,6 @@
 package org.apache.inlong.manager.client;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.compress.utils.Lists;
 import org.apache.inlong.manager.client.api.ClientConfiguration;
 import org.apache.inlong.manager.client.api.DataFormat;
 import org.apache.inlong.manager.client.api.DataSeparator;
@@ -30,25 +29,21 @@ import org.apache.inlong.manager.client.api.InlongGroupContext;
 import org.apache.inlong.manager.client.api.InlongStreamBuilder;
 import org.apache.inlong.manager.client.api.InlongStreamConf;
 import org.apache.inlong.manager.client.api.PulsarBaseConf;
-import org.apache.inlong.manager.client.api.SinkField;
-import org.apache.inlong.manager.client.api.StreamField;
 import org.apache.inlong.manager.client.api.auth.DefaultAuthentication;
-import org.apache.inlong.manager.client.api.sink.HiveSink;
-import org.apache.inlong.manager.client.api.source.KafkaSource;
-import org.apache.inlong.manager.common.enums.FieldType;
-import org.apache.inlong.manager.common.enums.FileFormat;
+import org.apache.inlong.manager.client.api.sink.KafkaSink;
+import org.apache.inlong.manager.client.api.source.MySQLBinlogSource;
+import org.apache.inlong.manager.common.enums.MQType;
 import org.apache.shiro.util.Assert;
 import org.junit.Test;
 
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 @Slf4j
-public class Kafka2HiveTest {
+public class Binlog2KafkaExample {
 
     // Manager web url
     public static String SERVICE_URL = "127.0.0.1:8083";
@@ -70,7 +65,7 @@ public class Kafka2HiveTest {
     public static String topic = "{pulsar.topic}";
 
     @Test
-    public void testCreateGroupForHive() {
+    public void testCreateGroupForKafka() {
         ClientConfiguration configuration = new ClientConfiguration();
         configuration.setWriteTimeout(10);
         configuration.setReadTimeout(10);
@@ -83,9 +78,8 @@ public class Kafka2HiveTest {
             InlongGroup group = inlongClient.forGroup(groupConf);
             InlongStreamConf streamConf = createStreamConf();
             InlongStreamBuilder streamBuilder = group.createStream(streamConf);
-            streamBuilder.fields(createStreamFields());
-            streamBuilder.source(createKafkaSource());
-            streamBuilder.sink(createHiveSink());
+            streamBuilder.source(createMysqlSource());
+            streamBuilder.sink(createKafkaSink());
             streamBuilder.initOrUpdate();
             // start group
             InlongGroupContext inlongGroupContext = group.init();
@@ -107,7 +101,45 @@ public class Kafka2HiveTest {
         InlongGroupConf groupConf = createGroupConf();
         try {
             InlongGroup group = inlongClient.forGroup(groupConf);
-            InlongGroupContext groupContext = group.delete();
+            InlongGroupContext groupContext = group.delete(true);
+            Assert.notNull(groupContext);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Test
+    public void testRestartGroup() {
+        ClientConfiguration configuration = new ClientConfiguration();
+        configuration.setWriteTimeout(10);
+        configuration.setReadTimeout(10);
+        configuration.setConnectTimeout(10);
+        configuration.setTimeUnit(TimeUnit.SECONDS);
+        configuration.setAuthentication(INLONG_AUTH);
+        InlongClient inlongClient = InlongClient.create(SERVICE_URL, configuration);
+        InlongGroupConf groupConf = createGroupConf();
+        try {
+            InlongGroup group = inlongClient.forGroup(groupConf);
+            InlongGroupContext groupContext = group.restart(true);
+            Assert.notNull(groupContext);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Test
+    public void testSuspendGroup() {
+        ClientConfiguration configuration = new ClientConfiguration();
+        configuration.setWriteTimeout(10);
+        configuration.setReadTimeout(10);
+        configuration.setConnectTimeout(10);
+        configuration.setTimeUnit(TimeUnit.SECONDS);
+        configuration.setAuthentication(INLONG_AUTH);
+        InlongClient inlongClient = InlongClient.create(SERVICE_URL, configuration);
+        InlongGroupConf groupConf = createGroupConf();
+        try {
+            InlongGroup group = inlongClient.forGroup(groupConf);
+            InlongGroupContext groupContext = group.suspend(true);
             Assert.notNull(groupContext);
         } catch (Exception e) {
             e.printStackTrace();
@@ -121,13 +153,13 @@ public class Kafka2HiveTest {
         inlongGroupConf.setProxyClusterId(1);
         //pulsar conf
         PulsarBaseConf pulsarBaseConf = new PulsarBaseConf();
+        pulsarBaseConf.setType(MQType.PULSAR);
         inlongGroupConf.setMqBaseConf(pulsarBaseConf);
         pulsarBaseConf.setPulsarServiceUrl(PULSAR_SERVICE_URL);
         pulsarBaseConf.setPulsarAdminUrl(PULSAR_ADMIN_URL);
         pulsarBaseConf.setNamespace("public");
         pulsarBaseConf.setEnableCreateResource(false);
         pulsarBaseConf.setTenant(tenant);
-
         //flink conf
         FlinkSortBaseConf sortBaseConf = new FlinkSortBaseConf();
         inlongGroupConf.setSortBaseConf(sortBaseConf);
@@ -154,40 +186,27 @@ public class Kafka2HiveTest {
         return streamConf;
     }
 
-    public KafkaSource createKafkaSource() {
-        KafkaSource kafkaSource = new KafkaSource();
-        kafkaSource.setBootstrapServers("{kafka.bootstrap}");
-        kafkaSource.setTopic("{kafka.topic}");
-        kafkaSource.setSourceName("{kafka.source.name}");
-        kafkaSource.setDataFormat(DataFormat.JSON);
-        return kafkaSource;
+    public MySQLBinlogSource createMysqlSource() {
+        MySQLBinlogSource mySQLBinlogSource = new MySQLBinlogSource();
+        mySQLBinlogSource.setDbNames(Arrays.asList("{db.name}"));
+        mySQLBinlogSource.setHostname("{db.url}");
+        mySQLBinlogSource.setAuthentication(new DefaultAuthentication("root", "inlong"));
+        mySQLBinlogSource.setSourceName("{mysql.source.name}");
+        mySQLBinlogSource.setAllMigration(true);
+        return mySQLBinlogSource;
     }
 
-    private HiveSink createHiveSink() {
-        HiveSink hiveSink = new HiveSink();
-        hiveSink.setDbName("{db.name}");
-        hiveSink.setJdbcUrl("jdbc:hive2://{ip:port}");
-        hiveSink.setAuthentication(new DefaultAuthentication("hive", "hive"));
-        hiveSink.setCharset(StandardCharsets.UTF_8);
-        hiveSink.setFileFormat(FileFormat.TextFile);
-        hiveSink.setDataSeparator(DataSeparator.VERTICAL_BAR);
-        hiveSink.setDataPath("hdfs://{ip:port}/usr/hive/warehouse/{db.name}");
-
-        List<SinkField> fields = new ArrayList<>();
-        SinkField field1 = new SinkField(0, FieldType.INT, "age", FieldType.INT, "age");
-        SinkField field2 = new SinkField(1, FieldType.STRING, "name", FieldType.STRING, "name");
-        fields.add(field1);
-        fields.add(field2);
-        hiveSink.setSinkFields(fields);
-        hiveSink.setTableName("{table.name}");
-        hiveSink.setSinkName("{hive.sink.name}");
-        return hiveSink;
-    }
-
-    public List<StreamField> createStreamFields() {
-        List<StreamField> streamFieldList = Lists.newArrayList();
-        streamFieldList.add(new StreamField(0, FieldType.STRING, "name", null, null));
-        streamFieldList.add(new StreamField(1, FieldType.INT, "age", null, null));
-        return streamFieldList;
+    private KafkaSink createKafkaSink() {
+        KafkaSink kafkaSink = new KafkaSink();
+        kafkaSink.setDataFormat(DataFormat.CANAL);
+        kafkaSink.setAddress("{kafka.bootstrap}");
+        kafkaSink.setTopicName("{kafka.topic}");
+        kafkaSink.setNeedCreated(false);
+        kafkaSink.setSinkName("{kafka.sink.name}");
+        Map<String, Object> properties = new HashMap<>();
+        //Not needed if kafka cluster is not set
+        properties.put("transaction.timeout.ms", 9000000);
+        kafkaSink.setProperties(properties);
+        return kafkaSink;
     }
 }
