+++ inlong-manager/pom.xml
@@ -39,7 +39,7 @@
         <module>manager-docker</module>
         <module>manager-plugin-examples</module>
         <module>manager-client</module>
-        <module>manager-client-test</module>
+        <module>manager-client-example</module>
     </modules>
 
     <properties>
